backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
torch>=2.3.0
torchvision>=0.18.0
pillow>=10.3.0
opencv-python>=4.10.0.84
numpy>=1.26.4
pydantic>=2.7.0
